Added public properties that tell you the color of the starting and ending rooms generated by the algorithm
So that we know where to put the white pieces at the beginning and where to put the stairs for the ending.

This makes me think about the risk of the team of white chess pieces not fitting into the randomly selected first room? (I guess if the minimum room is 3x3 tiles then it's okay... be funny to see them all cramped up in there...)