Added basics of chess notation to the game
# Notation style

Basically it's what you would expect, with the piece symbol (unless a pawn) followed by the destination square followed by capture information or check information. And then the flavour message if relevant.

Because the board is super wide there are more files than just a-h and indeed more than even a-z. For now I went with doubling letters beyond z, so you go from x y z to aa bb cc and so on. That gives us 52 files which is more than enough.

# X-Ray vision

A thing I hadn't thought about until I played with this version is that you get a little hint about what black is doing in the dark! This meant, for instance I made a move, saw a bishop had moved and could ascertain that it was near to my pieces. The very next move it captured a piece of mine!

Raises the question of whether the rank and file numbers ought to be visible on the screen as they often are in chess programs? This would make it plausible to track down black's last move more effectively which could be satisfying?