Updated notation code to include Rogue-like messages in a parentheticals
# Messages

Rogue doesn't seem(?) to care about the nature of the hit - an 'excellent' hit can take just one HP for example (or none?!). As such I just divided Rogue's messages into "hits" and "misses" and apply one or the other depending on whether the damage of the move was zero or above zero. It works fine.

Issue: "scored an excellent hit on" is long enough that it overruns the status line if white and black score an excellent his in the same move. Unsure of what to do about it other than eliminate that message, which is a shame because it's iconic. The other option is to divide black and white moves on the status line so you only see one, a pause, then the response. But the result of that will almost certainly be white missing out on seeing their own moves and it will remove some sense of the moves being in response to one another. Smaller font size is the only other alternative, which seems a shame as it's nice right now.

Missing: We'll need a notation for pawn promotion. We assume it's always a queen which is totally legit, but therefore the notation should be like  e21=Q. Similar to check, so can be handled in the same way. Even if it's like 1. d20xe21=Q+ (Your pawn defeated the black knight) < a pretty unlikely move, but nonetheless...