Fixed hallways drawing alongside rooms and blurring the distinction between room and hall.
# Blurry room-halls?

In the prior version it was possible for the pixels of a hallway in the generator to be right up against the pixels of a room. This is fine in terms of the level being navigable and even leads to some somewhat more interesting architectures because you get irregularly shapes rooms out of it. However we both felt that it conflicted too much with the visible algorithm which says that entering a room makes the whole room visible while being in a hall only reveals the squares around you, as per Rogue's rules. With the hallways touching the rooms it mean you could _feel_ like you were in a single room but sometimes you would reveal the whole thing (because you were on a tile that was secretly REALLY in the room) and sometimes you wouldn't reveal as much (because you were on a tile that was secretly REALLY a hallway). This fix removes that inconsistency and makes the game feel more successfully Rogue. Not to mention that Rogue's room architecture is boring anyway, they're all squares, it's fine.